requirements.txt
-f /usr/share/pip-wheels
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
djangorestframework==3.8.0
idna==2.6
pytz==2018.3
requests==2.18.4
urllib3==1.22
